requirements.txt
Django==1.10.2
django-bootstrap-dynamic-formsets==0.5.0
django-bootstrap3==7.1.0
django-ckeditor==5.1.1
django-crispy-forms==1.6.0
django-money==0.9.1
django-mptt==0.8.6
py-moneyed==0.6.0
